separate components from views
